Embed version string in binaries via Git
Use 'git describe --tags --always --dirty' to generate a version string
which gets included in every binary. The --tags is necessary, because
the latest tag 'v7.0' is not an annotated tag.

If the git-describe-output changes, make will rebuild the required files
automatically. Unfortunately, this is broken on Windows due to a bug in
Microchip's toolchain: https://www.microchip.com/forums/m1061336.aspx

Until that is fixed, a full rebuild is always done on Windows platforms,
because builds that are a bit slower are better than hard to debug
situations due to inconsistent builds.